Back-End/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
motor==3.7.0
pydantic==2.10.6
python-dotenv==1.0.1
passlib==1.7.4
numpy==1.26.1
torch==2.6.0
torchvision==0.21.0
pillow==10.1.0
slowapi==0.1.9